Use correct Spanish translation, refactor, update test
