+++ locales/active.es.toml
@@ -1,20 +1,20 @@
-btnCompletePayment = "Completar pagamente por registrante"
-btnContinueToCheckout = "Continuar a checkout"
-btnSendVerifEmail = "Enviar correo electronico por verification"
-errProcessPayment = "Era no possible pagar"
-frmAmount = "La Cuenta: {{ .CostPrint }} {{ .Currency }}"
-frmCity = "Citia"
+btnCompletePayment = "Completa el pago para registrarte"
+btnContinueToCheckout = "Continuar con finalizar el pago"
+btnSendVerifEmail = "Enviar correo electrónico de verificación"
+errProcessPayment = "No se pudo procesar el pago"
+frmAmount = "Monto {{.CostPrint}} {{.Currency}}"
+frmCity = "Ciudad"
 frmCountry = "País"
-frmEnterEmail = "Su addresso correo electronico por favor"
-frmFirstName = "Nombre primero"
-frmILiveIn = "Vivo en..."
-frmPaymentDetails = "Informacion de pagamente"
-frmSameEmail = "Address correo electronico - usar lo mismo"
-frmYourDetails = "Su informacion"
-pgCheckEmail = "consultar su correo electrónico, y haga clic en el enlace que hemos enviado"
-pgNowRegistered = "Ahora estas registrado!"
-pgRegisterFor = "Registrarse en {{ .Name }}"
-pgRegisteredFor = "Estás registrado para {{ .Name }}"
-valEnterEmail = "Por favor utilice una dirección de correo electrónico válida."
-valFirstName = "Se requiere un primer nombre válido"
-valSameEmail = "Por favor utilice una dirección de correo electrónico válida."
+frmEnterEmail = "Por favor ingrese su dirección de correo electrónico"
+frmFirstName = "Nombre"
+frmILiveIn = "Vivo en ..."
+frmPaymentDetails = "Detalles de pago"
+frmSameEmail = "Correo electrónico: use el mismo que verificó"
+frmYourDetails = "Sus datos"
+pgCheckEmail = "Por favor verifica tu bandeja de entrada de correo electrónico y haz clic en el enlace que te hemos enviado"
+pgNowRegistered = "¡Ya estás registrado!"
+pgRegisterFor = "Registrarse para {{.Nombre}}"
+pgRegisteredFor = "Registrado para {{.Nombre}}"
+valEnterEmail = "Ingresa una dirección de correo electrónico válida para que podamos enviarte detalles de la convención"
+valFirstName = "Se requiere un nombre válido."
+valSameEmail = "Ingresa una dirección de correo electrónico válida para que podamos enviarte detalles de la convención"
